refactor(layout): move synth/ to layers/generative/synth/ (T124.5.2)
