+++ docs/plan.md
@@ -3723,7 +3723,7 @@ left behind.
 - [x] T124.5.1 Move `gnn/` -> `layers/gnn/`  Owner: TBD  Est: 1h  verifies: [infrastructure]  DONE 2026-04-27
   4 files; literal GNN layers.
 
-- [ ] T124.5.2 Move `synth/` -> `layers/generative/synth/` (or experimental/)  Owner: TBD  Est: 1h  verifies: [infrastructure]
+- [x] T124.5.2 Move `synth/` -> `layers/generative/synth/` (or experimental/)  Owner: TBD  Est: 1h  verifies: [infrastructure] DONE 2026-04-27
   5 files; VAE-based synthetic data generation.
 
 - [ ] T124.5.3 Rename + relocate `shared/` -> `layers/shared_latent/`  Owner: TBD  Est: 1h  verifies: [infrastructure]
@@ -3828,7 +3828,7 @@ these fail in worktree isolation, drop to 4 per wave.)
 #### Wave E124-3: Tail consolidation + activation unification (8 agents)
 - [ ] T124.4.5 provenance/ -> training/provenance/  verifies: [infrastructure]
 - [ ] T124.4.6 federated/ -> training/federated/  verifies: [infrastructure]
-- [ ] T124.5.2 synth/ -> layers/generative/synth/  verifies: [infrastructure]
+- [x] T124.5.2 synth/ -> layers/generative/synth/  verifies: [infrastructure] DONE 2026-04-27
 - [ ] T124.5.3 shared/ -> layers/shared_latent/  verifies: [infrastructure]
 - [ ] T124.5.4 causal+features+regime -> inference/timeseries/  verifies: [infrastructure]
 - [ ] T124.5.5 modelcache+modeldsl+registry -> model/  verifies: [infrastructure]
