+++ layers/generative/synth/crash.go
