+++ layers/generative/synth/vae_test.go
