+++ tests/parity/layer_parity_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/zerfoo/zerfoo/layers/embeddings"
 	"github.com/zerfoo/zerfoo/layers/functional"
 	"github.com/zerfoo/zerfoo/layers/gather"
+	"github.com/zerfoo/zerfoo/layers/generative/synth"
 	"github.com/zerfoo/zerfoo/layers/gnn"
 	"github.com/zerfoo/zerfoo/layers/hrm"
 	"github.com/zerfoo/zerfoo/layers/normalization"
@@ -30,7 +31,6 @@ import (
 	"github.com/zerfoo/zerfoo/layers/transformer"
 	ltranspose "github.com/zerfoo/zerfoo/layers/transpose"
 	"github.com/zerfoo/zerfoo/layers/vision"
-	"github.com/zerfoo/zerfoo/synth"
 	"github.com/zerfoo/zerfoo/tabular"
 	"github.com/zerfoo/zerfoo/training/loss"
 	"github.com/zerfoo/zerfoo/training/optimizer"
